Simplified part-of-speech tags with parse(tagset=UNIVERSAL)
By default, the parsers in Pattern use the Penn Treebank tagset, which was developed for English. Pattern used to support only English, but multilingual support is growing. It is sometimes difficult to map tagsets of other languages to PTB. With parse(tagset=UNIVERSAL), simplified universal tags shared by all languages, are returned, loosely following the outlines in the paper by Petrov (2012).