Add package installation subprocess in setup
Even if the setup process completed without any problem, I had to install all the missing packages one by one — for this reason, I propose a simple way to automate this phase with the `subprocess module`. The advantage of `call()` vs `os.system()` is that it is more flexible, enabling you to get the stdout, stderr, status code, better error handling and so on.